[Refactor] commonMissionRepository분리
