+++ src/main/java/com/firefighter/aenitto/missions/repository/CommonMissionRepository.java
@@ -1,0 +1,11 @@
+package com.firefighter.aenitto.missions.repository;
+
+import com.firefighter.aenitto.missions.domain.CommonMission;
+
+import java.time.LocalDate;
+import java.util.Optional;
+
+public interface CommonMissionRepository {
+    CommonMission saveCommonMission(CommonMission commonMission);
+    Optional<CommonMission> findCommonMissionByDate(LocalDate date);
+}
