+++ src/main/java/com/firefighter/aenitto/missions/repository/MissionRepository.java
@@ -10,10 +10,7 @@ import java.time.LocalDate;
 import java.util.Optional;
 
 public interface MissionRepository {
-    CommonMission saveCommonMission(CommonMission commonMission);
+
     Optional<Mission> findRandomMission(MissionType missionType);
-
-    Optional<CommonMission> findCommonMissionByDate(LocalDate date);
-
     Optional<IndividualMission> findIndividualMissionByDate(LocalDate date, Long memberRoomId);
 }
